Refactor + rename configure -> parse
Move some more specific code for transforming pipeline and file groups
into main

Split out code that parses values + pipelines into full, final name TBD

Move MonifyGroup into main
