+++ parse/plugin.go
@@ -1,4 +1,4 @@
-package configure
+package parse
 
 import (
 	"encoding/json"
