+++ parse/plugin_test.go
@@ -1,4 +1,4 @@
-package configure
+package parse
 
 import (
 	"testing"
